[PT Run] Reversed refactoring NativeMethods.cs into single file
