+++ .github/actions/spell-check/expect.txt
@@ -9,7 +9,6 @@ Abug
 accctrl
 Acceleratorkeys
 ACCEPTFILES
-ACCESSDENIED
 accessibile
 accessibilityinsights
 Acl
