+++ src/modules/launcher/Plugins/Community.PowerToys.Run.Plugin.WebSearch/Community.PowerToys.Run.Plugin.WebSearch.csproj
@@ -104,9 +104,6 @@
   </ItemGroup>
   
   <ItemGroup>
-    <Compile Include="..\Common\NativeMethods.cs">
-      <Link>NativeMethods.cs</Link>
-    </Compile>
     <Compile Include="..\Common\BrowserInfo.cs">
       <Link>BrowserInfo.cs</Link>
     </Compile>
