+++ src/modules/launcher/Plugins/Microsoft.Plugin.Program/Programs/UWP.cs
@@ -11,6 +11,7 @@ using System.Runtime.InteropServices;
 using System.Runtime.InteropServices.ComTypes;
 using System.Xml.Linq;
 using Microsoft.Plugin.Program.Logger;
+using Microsoft.Plugin.Program.Win32;
 using Wox.Plugin.Logger;
 
 namespace Microsoft.Plugin.Program.Programs
@@ -62,10 +63,10 @@ namespace Microsoft.Plugin.Program.Programs
             InitPackageVersion(namespaces);
 
             const uint noAttribute = 0x80;
-            const NativeMethods.Stgm exclusiveRead = NativeMethods.Stgm.Read;
+            const Stgm exclusiveRead = Stgm.Read;
             var hResult = NativeMethods.SHCreateStreamOnFileEx(path, exclusiveRead, noAttribute, false, null, out IStream stream);
 
-            if (hResult == NativeMethods.HRESULT.S_OK)
+            if (hResult == Hresult.Ok)
             {
                 Apps = AppxPackageHelper.GetAppsFromManifest(stream).Select(appInManifest => new UWPApplication(appInManifest, this)).Where(a =>
                 {
@@ -212,6 +213,13 @@ namespace Microsoft.Plugin.Program.Programs
             Windows81,
             Windows8,
             Unknown,
+        }
+
+        [Flags]
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Naming", "CA1714:Flags enums should have plural names", Justification = "This name is consistent with the corresponding win32 flags: https://docs.microsoft.com/en-us/windows/win32/stg/stgm-constants ")]
+        public enum Stgm : long
+        {
+            Read = 0x00000000L,
         }
 
         public enum Hresult : int
